Metadata generation/modification/encryption works

+++ mb/drm_audio_fw/src/constants.h
@@ -202,7 +202,7 @@ typedef struct {
 // store of internal state
 typedef struct {
     char logged_in;             // whether or not a user is logged on
-    u8 uid;                     // logged on user id
+    u32 uid;                     // logged on user id
     char username[USERNAME_SZ]; // logged on username
     char pin[MAX_PIN_SZ];       // logged on pin
     song_md song_md;            // current song metadata
